Revert hack for musl libc
Decreasing the alien stack ptr by a word is expressly the wrong thing
because we want to catch, not hide, any mis-use of the stack.

So put in a diffent hack to detect structure size mismatch before
cold-init, and also have grovel-headers output a hardcoded definition
(for lack of a better idea) if the auto-defined struct would be wrong.