Fix bogus file-position calculation in READ-LINKAGE-CELLS
If the header didn't start at byte 0 of the core file, everything read in was
totally wrong. The arbitrary garbage bytes comprising the linkage space were
usually ignored by the "(not (in-bounds-p value text-bounds))" test in
bypass-indirection-cells. However, any word that randomly happened to look
in-bounds would be patched in, corrupting the containing function.