Allow absence of compiled-debug-info in code-debug-info slot
This facilitates implementing #+linkage-space without executable-funinstances,
as planned for ppc64. The wrapper on a closure installed as a globally named
function can be a simple-fun. Prior to git rev 2a615f1e the x86-64 logic for
"simplifying" a closure was similar but weirder - a code object with no
simple-fun which should not exist outside of the asm routines.