Deal with spurious tagbody tags in macroexpand-all
Given (defmacro foo () 'foo) then the expansion of (tagbody (go foo)
(foo) foo) will be (tagbody (go foo) foo foo), which is not right.
This tries to deal with that by checking the expansion of a form in
the body of a tagbody: if it looks like a tag it wraps it in progn.
So the expansion would now be (tagbody (go foo) (progn foo) foo).

I don't know if the use of relist is correct here.