Simplify autogenerated WITH-<c-struct> macro in sb-grovel
It shouldn't use UNWIND-PROTECT or call malloc, but instead just have WITH-ALIEN
utilize the alien stack. Nor should it gensym so much, or use so much EVAL-WHEN.
It's as if sb-grovel's author didn't know how macros or SB-ALIEN work.

In making this change I noticed that the field initializer list is always null
as used within SBCL. And as sb-grovel has no tests, it is impossible to conclude
that this aspect of the macro is even necessary, never mind bug-free.