Optimize SLOT-VALUE on structures within methods
CAN-OPTIMIZE-ACCESS has historically denied that slot access on
structure objects was possible, rendering some downstream code dead.
However, there doesn't seem to be a reason that we shouldn't optimize
SLOT-VALUE with known slot-name on structures within methods, so allow
them through and do the right thing based on the information in the
effective slot definition for the slot (including handling BOUNDP and
MAKUNBOUND appropriately).