Remove some build-time conditionalization for #+mark-region
In theory it should be possible to produce FASLs for either GC from an image
built with the other GC. Most of the problems in so doing stem from hardwired
static symbol addresses. With care, I think we can make them come out the same.
For functions that need no static symbols, this does seem to work as intended.
But of course it's very difficult never to need NIL in a function.