Slightly improve BACKTRACE-ALL-THREADS for x86-64
It no longer crashes without libunwind.
This is marginally better than code that users often craft resembing
 (DOLIST (THREAD (LIST-ALL-THREADS)) (INTERRUPT-THREAD THREAD #'PRINT-BACKTRACE))

Among the problems with interrupt-thread are that backtrace may call dladdr()
which is not usually signal-safe (who knows?) and therefore is prone to deadlock.

By hooking into stop-the-world, we are somewhat confident that any thread stopped
in Lisp gets a reasonable backtrace, and that if dladdr() is called, the call
occurs not in a signal handler but rather the foreground thread. This has a
fixable bug that some stopped thread may still hold a needed lock.