Fix CLISP warnings about the initialzation of COMPONENT slot OUTER-LOOP.
The initform for slot OUTER-LOOP of structure COMPONENT is:

    (make-loop :kind :outer :head head :tail (list tail))

CLISP outputs warnings about treating HEAD and TAIL as special variables
because they are unbound.

ANSI spec section 8.1.1 says the following about structure slot initialization
forms:

    The slot default init forms are evaluated in the lexical environment in
    which the defstruct form itself appears and in the dynamic environment in
    which the call to the constructor function appears.

The CLISP warning is legit.  HEAD and TAIL are unbound in the lexical
environment in which the COMPONENT defstruct form appears.  HEAD and TAIL are
parameters of the constructor function for COMPONENT structs.  They should be
unbound in the lexical environment in which the form is evaluated.

Why does the original code work?