Support typed circularities in the #S reader
Use the escape hatch to allow through invalid values for structure
slots if they contain the reader circularity markers when calling the
default constructor from #S, and if the slot is not a raw slot.

When traversing the resulting structure, provide a delayed typecheck
function to perform the type check after all the circularity markers
have been substituted.