Shrink the size of callback assembler wrappers on many platforms.
On some targets with callback_wrapper_trampoline, we were redundantly
computing and passing in ENTER-ALIEN-CALLBACK. We don't need to do
that, and to make things more similar between threaded and
non-threaded builds, create a callback_wrapper_trampoline for
non-threaded builds as well that handles computing
ENTER-ALIEN-CALLBACK. This has the benefit of shrinking the size of
the callback wrapper almost everywhere else on the assembly side, even
if the C function still has to compute it. This also reduces some
feature conditionals and hopefully should make it marginally easier to
improve or port the callback mechanism in the future.

Doing this is also a requirement for callbacks to work with
relocatable static space, since hardwired static space addresses in
assembly won't work after static space is relocated.