Remove :set-trans for %symbol-global-value
Each backend defines a %SET-SYMBOL-GLOBAL-VALUE vop that is directly chosen
by name in IR2-CONVERT-SET. Nothing is gained by a :set-trans, and in fact
the custom vop is marginally better because it might accept an immediate SC
for the symbol, not that any of the lines changed herein benefit from it.
One fewer special case in x86-64 SET-SLOT is good in any event.