Fix massively wrong tune_asm_routines_for_microarch
By writing a byte into a completely arbitrary instruction, this did not cause
visible harm but invisibly disabled two runtime optimizations. Small miracle.

The asm code is trying to achieve the following: if and only if FILL is large
enough, and the CPU says that STOSQ is faster than a loop, then use STOSQ.
The bad poke affected a comparison of the fill element versus NIL to see if
the GC barrier can be entirely skipped, which it never was. In addition,
the optimized path for STOSQ was unreachable.

This all happened with changing to soft card marks (rev 9394492f).

I'd like to redo the "tuning" to occur in Lisp, but meanwhile this patch adds
safeguards against storing to random places.