Change many uses of NIL-VALUE to a NULL-TN symbol macro
This does not actually implement NIL in a register but merely gives a different
name to NIL as a machine operand. I contemplated renaming to NIL-OPERAND
to generalize over NIL-VALUE or NULL-TN but decided against doing an extra
step. The overall intent is that the rename is ignorable, but makes the
patch which truly implements NULL-TN this much less to look at.