Fix elfcore test
The heap relocation subtest has been using the wrong fuzzing data file for
the past 2 years or more. Consequently for lack of a third column of numbers
in the test inputs, it has not been doing anything as directed.

The test should be strengthened to assert that it isn't completely random,
but I don't know how, since the premise is that mmap might not do what you
want, and so the test may fail to fake the desired mmap failure.