Delete unbound-marker-bits macro
The problem allegedly solved by having MAKE-UNBOUND-MARKER return a word
pointing "somewhere" was so that a (misaligned) dereference does not crash.
But enough things are different that the original patch is now irrelevant.
Orginally this was implemented so that EQUAL and EQUALP hash-tables could
test any probed key against the supplied key without checking for any empty
slot. But in any case, EQUAL and EQUALP seem fine with a unbound-marker.