Remove comment and use non-macro constructors.
The FIXME for MAKE-BPT-LRA is misleading, as the primary value of
MAKE-BPT-LRA is in fact a freshly consed LRA. The fact that a
containing code object was created which the LRA is embedded in is a
secondary fact.

Also change some gratuitous usages of MACROLET to usages of lexical
variables and inline functions.