x86-64: Order core directory spaces more favorably
I'm seeing FIXEDOBJ space often get ENOMEM which should never happen.
I think it has to do with containerized executions not being able to search
for a hole. If nothing else, this outputs the contents of /proc/self/maps
which was done only for immovable mappings. Since all mappings are movable
for x86-64 now, the debugging output was never shown.