Remove super weird defconstant
There was no need to define it in one file but use it in another, or was there?
Well, apparently an excruciatingly subtle reason: If cold-init's LTV fixup
hadn't gotten to that constant, then it would patch in 0, and attempting to
iterate over the vector would crash even though genesis actually did assign a
value. This makes no sense at all. Just use the value that genesis dumped.