Fix auto stack allocation in the presence of local calls.
The test case demonstrates that we need to be careful about multiple
xep references hidden by local calls, so the condition for when to
conitionally stack allocate gets stricter. In fact the problem is the
same as with several external functional references in terms of where
to insert cleanup code.

Reported by Andreas Franke.