Don't mark local functions DX just because the ENCLOSE is DX.
We were being too aggressive in downwards-dxifying. As the new test
case shows, just because one local function is DX doesn't mean that
downward dxifying should immediately make the other local function
stack allocatable. Just mark the REF argument lvar as potentially
stack-allocatable so that the usual find-stack-allocatable-parts
machinery can apply and actually make the decision whether the
function is safe to stack allocate.

Also fix one of the other test cases so that it doesn't cons outside
of the closure consing.