stack: Don't preserve intervening stack pointers too pessimistically
It's actually fine to walk up the cleanup nesting from the current
node being stack-mess-up-walked, as long as we don't do it for dynamic
extent nodes. And we don't need to do it for dynamic extent nodes
because no stack allocation is actually happening there. This has the
effect of not being overly pessimistic. Add some additional test cases
just to be on the safe side.