sb-simd: Portability fixes
This is mostly just building on what already existed, primarily
only defining and calling VOPs where the instruction set exists.
Also a phony simd-pack-256 type to allow building when that
type doesn't exist.
And disable the vref vop implementation outside x86oids, as it's entirely
specific to that instruction set.