Style-warn if assoc/rassoc/member MPH transform fails unexpectedly
i.e. in situations where it appears - by guessing - that the user thought
the compiler capable of producing a minimal perfect hash in (SPEED 3),
but it did not do for whatever reason.
Though a fine line between an efficiency note vs warning, I think by making
it a warning, the user might get the hint that other Lisp compilers would
probably need it expressed as a hash-table lookup to begin with.