Optimize out a few instructions in perfect hash calculation
For sets of only character or only fixnum, there is no need to guard
against an object having the wrong type. So instead of doing
(if (fixnump x) (ldb ... x)) just use (ldb ... (get-lisp-obj-address x))
eliminating an initial branch. Also the AND operation - at least on
x86-64 - avoids use of a raw constant. This just changes some shifts
in the perfect hash expression. It drops a bit (or more) of precision,
but we're already doing that in descriptor-hash32 for the situations
where the keyset contains a mixture of immediates and symbols.