Use PROPER-LIST-P instead of (NULL (CDR (LAST x)))
And don't use ASSERT to mean ERROR. It was that way since rev 0a3d799b
but invariant checks should be done using the lighter-weight AVER. Except
this is a type-check of user inputs. Almost nowhere else are users allowed
to edit the supplied args (apart from all the CLHS-stipulated restarts
on package operations perhaps) to system function in this manner.