pcl: Revive compiler-less dfun code.
This helps applications which do not want to call the compiler at
runtime (perhaps allowing for applications which strip the compiler
entirely) and also allows flexibility of bootstrapping more pcl code
earlier.

Currently hidden behind an internal special variable which is always
nil, but could be part of a more general interface for controlling the
presence of the compiler.